refactor(raffle-instance): split monolithic lib.rs into domain submodules
- Extract 7 domain modules: admin, claim, draw, helpers, init, tickets, views
- lib.rs now contains only types, constants, and thin #[contractimpl] delegates
- Keep pre-existing events.rs and randomness.rs unchanged
- All 16 tests pass, clippy zero warnings
- Every new submodule is under 300 lines
