+++ contracts/raffle-instance/src/helpers.rs
@@ -1,0 +1,202 @@
+use soroban_sdk::{token, Address, BytesN, Env, Vec};
+
+use crate::events::{RaffleFinalized, RaffleStatusChanged, WinnerDrawn};
+use crate::randomness::{OracleSeedWinnerSelection, WinnerSelectionStrategy};
+use crate::{
+    DataKey, Error, FairnessMetadata, Raffle, RaffleStatus, RandomnessType, Ticket,
+};
+
+pub(crate) fn read_raffle(env: &Env) -> Result<Raffle, Error> {
+    env.storage().instance().get(&DataKey::Raffle).ok_or(Error::NotInitialized)
+}
+
+pub(crate) fn write_raffle(env: &Env, raffle: &Raffle) {
+    env.storage().instance().set(&DataKey::Raffle, raffle);
+}
+
+pub(crate) fn require_admin(env: &Env) -> Result<Address, Error> {
+    let admin: Address = env.storage().persistent().get(&DataKey::Admin).ok_or(Error::NotAuthorized)?;
+    admin.require_auth();
+    Ok(admin)
+}
+
+pub(crate) fn get_ticket_owner(env: &Env, ticket_id: u32) -> Option<Address> {
+    env.storage().persistent().get::<_, Ticket>(&DataKey::Ticket(ticket_id)).map(|t| t.owner)
+}
+
+pub(crate) fn acquire_guard(env: &Env) -> Result<(), Error> {
+    if env.storage().instance().has(&DataKey::ReentrancyGuard) {
+        return Err(Error::Reentrancy);
+    }
+    env.storage().instance().set(&DataKey::ReentrancyGuard, &true);
+    Ok(())
+}
+
+pub(crate) fn release_guard(env: &Env) {
+    env.storage().instance().remove(&DataKey::ReentrancyGuard);
+}
+
+pub(crate) struct Guard<'a> {
+    env: &'a Env,
+}
+
+impl<'a> Guard<'a> {
+    pub(crate) fn new(env: &'a Env) -> Result<Self, Error> {
+        acquire_guard(env)?;
+        Ok(Guard { env })
+    }
+}
+
+impl<'a> Drop for Guard<'a> {
+    fn drop(&mut self) {
+        release_guard(self.env);
+    }
+}
+
+#[allow(dead_code)]
+pub(crate) fn enforce_swap_guard(
+    env: &Env, raffle: &Raffle, amount_out: i128, min_amount_out: i128,
+) -> Result<(), Error> {
+    let deadline = env.ledger().timestamp() + raffle.swap_deadline_seconds;
+    if env.ledger().timestamp() > deadline {
+        return Err(Error::DeadlinePassed);
+    }
+    if amount_out < min_amount_out {
+        return Err(Error::SlippageExceeded);
+    }
+    Ok(())
+}
+
+pub(crate) fn request_randomness(env: &Env) -> Result<u64, Error> {
+    let already: bool = env.storage().instance().get(&DataKey::RandomnessRequested).unwrap_or(false);
+    if already { return Err(Error::RandomnessAlreadyRequested); }
+
+    use soroban_sdk::xdr::ToXdr;
+    let request_id_xdr = (
+        env.ledger().timestamp(),
+        env.ledger().sequence(),
+        env.current_contract_address().to_xdr(env),
+    ).to_xdr(env);
+    let request_id_hash: BytesN<32> = env.crypto().sha256(&request_id_xdr).into();
+    let arr = request_id_hash.to_array();
+    let mut id_bytes = [0u8; 8];
+    id_bytes.copy_from_slice(&arr[..8]);
+    let request_id = u64::from_be_bytes(id_bytes);
+
+    env.storage().instance().set(&DataKey::RandomnessRequested, &true);
+    env.storage().instance().set(&DataKey::RandomnessRequestLedger, &env.ledger().sequence());
+    env.storage().instance().set(&DataKey::RandomnessRequestId, &request_id);
+    Ok(request_id)
+}
+
+pub(crate) fn transition_to_drawing(env: &Env, raffle: &mut Raffle, timestamp: u64) -> Result<(), Error> {
+    let drawing_lock: bool = env.storage().instance().get(&DataKey::DrawingLock).unwrap_or(false);
+    if drawing_lock { return Err(Error::DrawingAlreadyInProgress); }
+
+    if raffle.status != RaffleStatus::Active {
+        if raffle.status == RaffleStatus::Drawing { return Err(Error::DrawingAlreadyInProgress); }
+        return Err(Error::InvalidStatusForDrawingTransition);
+    }
+
+    let old_status = raffle.status.clone();
+    raffle.status = RaffleStatus::Drawing;
+    write_raffle(env, raffle);
+    RaffleStatusChanged { old_status, new_status: RaffleStatus::Drawing, timestamp }.publish(env);
+    env.storage().instance().set(&DataKey::DrawingLock, &true);
+    Ok(())
+}
+
+pub(crate) fn require_not_paused(env: &Env) -> Result<(), Error> {
+    if env.storage().instance().get(&DataKey::Paused).unwrap_or(false) {
+        return Err(Error::ContractPaused);
+    }
+    Ok(())
+}
+
+pub(crate) fn validate_token_address(env: &Env, token_address: &Address) -> Result<(), Error> {
+    let token_client = token::Client::new(env, token_address);
+    let _ = token_client.try_decimals().map_err(|_| Error::InvalidTokenAddress)?;
+    Ok(())
+}
+
+pub(crate) fn build_internal_seed_u64(env: &Env) -> u64 {
+    use soroban_sdk::xdr::ToXdr;
+    let xdr = (
+        env.ledger().timestamp(),
+        env.ledger().sequence(),
+        env.current_contract_address(),
+    ).to_xdr(env);
+    let hash: BytesN<32> = env.crypto().sha256(&xdr).into();
+    let arr = hash.to_array();
+    let mut bytes = [0u8; 8];
+    bytes.copy_from_slice(&arr[..8]);
+    u64::from_be_bytes(bytes)
+}
+
+pub(crate) fn calculate_tier_prize(raffle: &Raffle, tier_index: u32) -> Result<i128, Error> {
+    let last_tier_index = raffle.prizes.len() - 1;
+    if tier_index == last_tier_index {
+        let mut allocated = 0i128;
+        for i in 0..last_tier_index {
+            let bp = raffle.prizes.get(i).unwrap();
+            let amt = raffle.prize_amount.checked_mul(bp as i128).ok_or(Error::ArithmeticOverflow)? / 10000;
+            allocated = allocated.checked_add(amt).ok_or(Error::ArithmeticOverflow)?;
+        }
+        return raffle.prize_amount.checked_sub(allocated).ok_or(Error::ArithmeticOverflow);
+    }
+    let bp = raffle.prizes.get(tier_index).unwrap();
+    raffle.prize_amount.checked_mul(bp as i128).ok_or(Error::ArithmeticOverflow).map(|a| a / 10000)
+}
+
+pub(crate) fn do_finalize_with_seed(
+    env: &Env, mut raffle: Raffle, seed: u64, randomness_type: RandomnessType,
+) -> Result<(), Error> {
+    let total_tickets = raffle.tickets_sold;
+    if total_tickets == 0 { return Err(Error::NoTicketsSold); }
+    if raffle.prizes.len() > total_tickets { return Err(Error::MorePrizesThanTickets); }
+    if raffle.tickets_sold == 0 { return Err(Error::NoActiveTickets); }
+
+    let selector = OracleSeedWinnerSelection::new(seed);
+    let winning_ticket_ids = selector.select_winner_indices(env, total_tickets, raffle.prizes.len());
+    let mut winners = Vec::new(env);
+
+    for i in 0..winning_ticket_ids.len() {
+        let idx = winning_ticket_ids.get(i).ok_or(Error::InvalidIndex)?;
+        let winner = get_ticket_owner(env, idx + 1).ok_or(Error::TicketNotFound)?;
+        winners.push_back(winner.clone());
+        WinnerDrawn { winner, ticket_id: idx, tier_index: i, timestamp: env.ledger().timestamp() }.publish(env);
+    }
+
+    let mut claimed_winners = Vec::new(env);
+    for _ in 0..raffle.prizes.len() { claimed_winners.push_back(false); }
+
+    env.storage().persistent().set(&DataKey::RandomnessSeed, &FairnessMetadata {
+        seed,
+        randomness_source: raffle.randomness_source.clone(),
+        winning_ticket_indices: winning_ticket_ids.clone(),
+        draw_timestamp: env.ledger().timestamp(),
+        draw_sequence: env.ledger().sequence(),
+    });
+
+    raffle.status = RaffleStatus::Finalized;
+    raffle.winners = winners.clone();
+    raffle.claimed_winners = claimed_winners;
+    raffle.finalized_at = Some(env.ledger().timestamp());
+    write_raffle(env, &raffle);
+
+    env.storage().instance().remove(&DataKey::RandomnessRequested);
+    env.storage().instance().remove(&DataKey::RandomnessRequestId);
+    env.storage().instance().remove(&DataKey::RandomnessRequestLedger);
+    env.storage().instance().set(&DataKey::DrawingLock, &false);
+
+    RaffleFinalized {
+        raffle_id: env.current_contract_address(),
+        winners, winning_ticket_ids,
+        total_tickets_sold: raffle.tickets_sold,
+        randomness_source: raffle.randomness_source.clone(),
+        randomness_type,
+        finalized_at: env.ledger().timestamp(),
+    }.publish(env);
+
+    Ok(())
+}
